server/requirements.txt
flask
tensorflow
openai
pillow
